{{cookiecutter.project_directory}}/pyproject.toml
[tool.poetry]
name = "Fastapi-Cookiecutter"
version = "0.1.0"
description = ""
authors = ["GGGoingdown <eddiepai0401@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.0"}
gunicorn = "^20.1.0"
tenacity = "^8.0.1"
dependency-injector = "^4.39.1"
aerich = "^0.6.3"
loguru = "^0.6.0"
sentry-sdk = "^1.5.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^3.2.0"
passlib = "^1.7.4"
email-validator = "^1.1.3"
pydantic = "^1.9.0"
python-multipart = "^0.0.5"
aioredis = "^2.0.1"
pre-commit = "^2.18.1"
shortuuid = "^1.0.8"
httpx = "^0.22.0"
celery = "^5.2.6"
redis = "^4.2.2"

[tool.poetry.dev-dependencies]
requests = "^2.27.1"
pytest = "^7.1.1"
pytest-asyncio = "^0.18.3"
pytest-ordering = "^0.6"


[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
